layout: Port layouter to use the sdapsarray environment directly
This simplifies the code somewhat as it does not need to care about
horizontal/vertical mode itself anymore. instead it can just rely on
the flipping feature of sdapsarray.

This commit also means that you can savely use fargile commands in
choice labels (questions are still pending) if you specify a text
in the optional argument.