base: vhbox now behaves more like line or paragraph depending on the mode.
If in horizontal mode, then no paragraph spacing is inserted. If in vertical
mode the paragraph spacing is inserted and everpar code is run. After the
environment the paragraph space is removed again, and is reinserted if \par
is encountered.

This way the textbox can in a way behave as part of the paragraph.