layout: Fix output of metadata at start of range array.
Some of the content in an sdapsarray may be discarded. The range
implementation put the question into a non-existing header cell
(by inserting a new line) so that it did not make it into the
metadata file.
The fix is to move the start of the question to be inside the
same box as the text itself.