array: Rework column and row spacing
For the column spacing we simply insert \tabcolsep on each side of every
cell for now. This nicely indents the table and gives us a configurable
separator for a start.

For the horizontal spacing the problem is a bit more complicated. Here
sdapsarray tries to behave as much as possible as normal content. It is
optimized for the row headers to contain multiple lines of text. If this
happens the following is done:
 * The first baseline is used to calculate the interline skip to the
   previous row
 * The vbox is re-packed so the depth of the last line is used
 * If the first and last baseline differ then the row header vbox is
   centered instead of being alinged to the baseline of the cells.

The exception to this rule is is the header row for which the last baseline
is used for all purposes.