base: Tweak the checkbox positioning some more
We now center it to 0.8ex as the capital letters are more like 1.6ex. The
result looks good with the default TeX font, it might need more tweaking
for other fonts.

Quite likely we will need to introduce an option to override the baseline
calculation if this turns out to be bad in other cases.