base: Allow choosing checkbox style and improve style code
We need two fundamental checkbox styles. That is "multichoice" and
"singlechoice" and allowing each to be set to different defaults. So
add \sdaps_checkbox_set_style:n and handle it so that the default form
changes from "box" to "ellipse" if it is set to "singlechoice".

This also changes the overrides and context behaviours where one now
needs to set the variables on "multichoice"/"singlechoice" instead of
"checkbox" (i.e. changing the size will require setting for both requires
setting it twice).