layout, classic: Allow singlechoice mode in choicequestion and choicearray
With this there is a relatively consistent way of switching multiple choice
questions to single choice questions. The main change here is that it is
now also permitted to assign a value to checkboxes so that these can be set
if used in single choice mode (as the assignment happens to the questions
variable rather than the checkboxes).