classic: use \verbatim@ instead of \verbatim@start
The @start macro simply checks whether the first line is empty and strips
it. For some reason this is not quite working here though, so instead of
calling \verbatim@start just call \verbatim@ directly which will cause the
items on the first line not to be discarded.

Note that the any whitespace to the first token will have been removed by
the code scanning for the optional parameter. This is likely the reason for
the issue to be appearing now in the first place.