base/layout: Rework variable handling completely
The variable handling had some major flaws causing issues for the
layouter. Rework the variable handling so that every leven of QObjects
will always get an implicit variable name. This implicit name will be
discarded if a later object (or answer) provides a variable name of
itself.
It is possible to prefix a variable name with "_" so that the implicit
variable names from the parent will not be discarded.