base: Make metadata writing a global variable bound to the context
Having a writepos setting on the boxes is not a good solution, as all
metadata handling needs to be disabled if not only the box position but
also other metadata needs to be supressed. Instead, make this a setting
on the context which will affect all metadata writing.

Note that metadata writing needs to be enabled for \sdaps_begin: and
\sdaps_end:.