array: Fix header rotation placement
The header is supposed to be placed centered with the baseline. Before
it was somewhat misplaced, also, to center it around the baseline we
need to take enough space to each side.

So, place it correctly, and take the maximum of height/depth space to
each side so that it is properly centered. In the extreme case, this
causes the insertion of a lot of whitespace, but we cannot prevent this
as we cannot shift the center for the cells underneath.