layout: Disable automatic alignment of similar environments
While this is a nice feature, it turns out that it is confusing more
often than it helps (in particular when alignment happens between
multiple sections of a document). It may make sense to advertise it more
prominently in the documentation, but we can safe users from unexpected
situations by disabling it by default.

Fixes #15