sphinx: Fix sdapsarray token rescan issue in array example
We have an issue where the first & character is read for the code that
checks the optional parameter. We are then re-configuring the parser,
but the token has already be read and will not be re-intepreted with the
new configuration. This causes an alignment character to be found in an
invalid location.

Fix this by inserting an empty group and add a comment about what is
going on.