array: Fix & not working correctly at start of environment
The code for optional argument parsing needs to peek the next character
in the stream which means that the catcode is resolved at that time. As
we override the category code for the & character, we might end up with
an alignment character rather than an active character with unexpected
results.

Fix this by using \peek_charcode_remove:NT to remove the & character
from the stream and emitting \sdaps_array_alignment: directly.

Thanks to Bruno Le Floch for suggesting this solution!