Fix Pages metadata generation in some corner cases
In certain corner cases, the Pages metadata may not have been written
out. To avoid this problem, write out the metadata after each page, thus
writing it earlier and ensuring it is not lost. Note that doing so is
fine as the main program will pick up the last entry in from the
metadata file.

Fixes: sdaps#192