classic: Improve spacing in multiline choicequestion items
The choicequestion items only had a \strut at the start, which was one
issue causing short line spacings to the next item in some cases. The
other thing we now do is correctly propagate the \prevdepth from the
last baseline in the choiceitem row, rather than the one relative to the
actual baseline that is used for aligning the items to each other.