layout: Remove grouping from array environments
As these environments do not actually print any headline or similar, it
makes sense to not emit grouping metadata. This should be handled by the
surrounding code.

Unfortunately, this is an API change, but very few users should exist of
this code. And the choicegroup/optiongroup/rangegroup environments will
still behave identical with the changes there.