base: Use question variable for single choice and limit overrides
Before singlechoice boxes would require both their own variable and the
value to override. As singlechoice boxes now simply use the variable of
their surrounding questions, it is sane to just only match against
variable + value.

TODO: This works now, but the whole thing of passing both variable and
value at the same time (and putting it into the metadata) is maybe a bit
odd?